requirements.txt
flask>=2.2.0
google-generativeai>=0.8.5
python-dotenv>=1.0.0
google-auth>=2.6.0